backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
pydantic==2.7.1
celery[redis]==5.4.0
redis==5.0.4
anthropic==0.28.0
voyageai==0.2.3
resend==0.7.2
httpx==0.27.0
psycopg2-binary==2.9.9
pgvector==0.3.2
playwright==1.44.0
python-multipart==0.0.9
svix==1.14.0
beautifulsoup4==4.12.3
lxml==5.2.2
tenacity==8.3.0
PyJWT[crypto]==2.8.0
reportlab==4.2.0
flower==2.0.1
sentry-sdk[fastapi]==1.45.0
slowapi==0.1.9
